Enum bin_prot::error::Error

pub enum Error {

    Io(Error),
    EndOfStream,
    SeqSizeNotProvided,
    MapSizeNotProvided,
    InvalidByte {
        byte: u8,
        dtype: String,
        allowed: Vec<u8>,
    },
    InvalidUtf8 {
        bytes: Vec<u8>,
    },
    InvalidOptionByte {
        got: u8,
    },
    InvalidIntegerByte {
        byte: u8,
    },
    VariantIndexTooLarge {
        index: u32,
    },
    DestinationIntegerOverflow,
    WontImplement,
    UnexpectedEndOfLayout,
    LayoutIteratorError,
    DeserializingLooseTypeWithoutLayout,
    UnknownCustomType {
        typ: String,
    },
    UnimplementedRule,
    UnknownPolyvarTag(u32),
    StreamNotFullyConsumed(usize),
    Custom {
        message: String,
    },

}

Error type for BinProt serialization and deserialization

Variants

Io(Error)

Error in underlying IO

EndOfStream

The end of stream was reached unexpectedly.

SeqSizeNotProvided

Size not given when serializing a sequence

MapSizeNotProvided

Size not given when serializing a sequence

InvalidByte

Fields

byte: u8

The byte that is invalid

dtype: String

The data type that the deserializer is attempting to deserialize

allowed: Vec<u8>

Bytes that are allowed in this context

Invalid byte encountered when deserializing

InvalidUtf8

Fields

bytes: Vec<u8>

Invalid byte sequence encountered

Invalid utf-8 char

InvalidOptionByte

Fields

got: u8

Invalid byte encountered

Invalid byte encountered deserializing option

InvalidIntegerByte

Fields

byte: u8

Invalid byte encountered

Invalid integer prefix byte

VariantIndexTooLarge

Fields

index: u32

Invalid index encountered

Encountered a variant index larger than allowed

DestinationIntegerOverflow

Destination integer type too small

WontImplement

Functionality will not be implemented. Probably it does not make sense for this format

UnexpectedEndOfLayout

The layout ended before the reader is empty

LayoutIteratorError

Layout iterator errored, can be due to invalid layout or mismatch between layout and input

DeserializingLooseTypeWithoutLayout

Have not provided a layout

UnknownCustomType

Fields

typ: String

The custom type identifier

An unknown custom type found in layout that deserializer code does not know not to handle

UnimplementedRule

There is no logic implemented to deserialize this rule yet

UnknownPolyvarTag(u32)

When deserializing a polyvar the tag does not match any known tags for the type

StreamNotFullyConsumed(usize)

When deserializing a polyvar the tag does not match any known tags for the type

Custom

Fields

message: String

The user-defined error message.

Some user-defined error occurred.