Module bin_prot::value

The Value enum, a loosely typed way of representing any valid bin_prot value.

Since bin_prot is not a self describing format, deserializing to a loosely typed value required a supplimentary file that describes the layout of the binary (see layout/)

Modules

layout

Defines a BinProt layout A layout is a data structure that defines a BinProt type. This is essentially moving the type information from compile time to runtime.

ser

Serialization implementation for a BinProt loosely typed Value

Structs

EnumData

This is the data that needs to be serialized and passed to the serde deserializer so that the value::Sum can know its variant name and index as described in the layout

Enums

Value

Recursive enum that can define any BinProt serializable type

Traits

Index

A type that can be used to index into a bin_prot::Value.