Module bin_prot::value::layout

Defines a BinProt layout A layout is a data structure that defines a BinProt type. This is essentially moving the type information from compile time to runtime.

Structs

HashTblEntry

Entry in a BinProt hash table hash types for the key and value

Layout

The main top level type for a layout file. Parse into this from json

RecordField

Field of a BinProt record with a name and a value

ResolvedPayload

Resolved reference of payload

Summand

Variant of a BinProt sum type (enum) has a name, index and zero-or-more BinProt values

TaggedPolyvar

τ ≤ Γ(a), τ is an instance of Γ(a) and (Γ(a) a type scheme

UnresolvedPayload

Unable to resolve reference of payload

Enums

BinProtRule

Recursively defined BinProtRule is how the type tree is constructed

Polyvar

τ ≤ Γ(a), τ is an instance of Γ(a) and (Γ(a) a type scheme

RuleRef

Define rule reference into Resolved and Unresolved Payload